Started refactoring code for explicit translation.
Moved the address offset calculation to a seperate function.
Planning todo the same with register and intermediate operands.
With these functions and the generate functions I want to move to more
explicit translating (e.g. Removing the translate_operand function and
its use with direct translatable instructions.)
This will expand the code size but give more control and options for
optimization.