Finished major refactoring round.
Changed the definition of reil_register, reil_integer and
reil_subaddress types.
Both reil_register and reil_integer now contain their sizes.
The reil_subaddress contains the offset of the corresponding x86
instruction and the reil instruction offset.
All the gen_* functions are changed to the consistent format which
includes the operand types in the names and the order of the operands
being input1, [input2], output.
Also added a lot of helper functions to reduced the duplicated code.
Fixed some bugs in the mean time as well.